requirements.txt
requests>=2.32.5
pycryptodome>=3.23.0
